requirements.txt
absl-py==0.7.1
aiocontextvars==0.2.2
alabaster==0.7.12
allennlp==0.9.0
asn1crypto==0.24.0
astor==0.8.0
attrs==19.3.0
Babel==2.8.0
bert-serving-client==1.10.0
bert-serving-server==1.10.0
blis==0.2.4
boto==2.49.0
boto3==1.11.11
botocore==1.14.11
bpemb==0.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
conllu==1.3.1
contextvars==2.4
cryptography==2.1.4
cycler==0.10.0
cymem==2.0.3
decorator==4.4.1
Deprecated==1.2.7
docutils==0.15.2
editdistance==0.5.3
filelock==3.0.12
flair==0.4.5
flaky==3.6.1
Flask==1.1.1
Flask-Cors==3.0.8
ftfy==5.6
future==0.18.2
gast==0.2.2
gensim==3.8.1
gevent==1.4.0
google-pasta==0.1.7
GPUtil==1.4.0
greenlet==0.4.15
grpcio==1.21.1
h5py==2.9.0
hyperopt==0.2.3
idna==2.6
imagesize==1.2.0
immutables==0.11
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.14.1
jsonnet==0.14.0
jsonpickle==1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
langcodes==1.4.1
langdetect==1.0.7
loguru==0.4.1
marisa-trie==0.7.5
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
more-itertools==8.2.0
mpld3==0.3
msgpack==0.6.2
murmurhash==1.0.2
networkx==2.2
nltk==3.4.5
numpy==1.16.4
numpydoc==0.9.2
overrides==2.8.0
packaging==20.1
pandas==1.0.1
parsimonious==0.8.1
plac==0.9.6
pluggy==0.13.1
preshed==2.0.1
protobuf==3.8.0
py==1.8.1
pycrypto==2.6.1
Pygments==2.5.2
pygobject==3.26.1
pyparsing==2.4.6
pytest==5.3.5
python-apt==1.6.4
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2019.3
pyxdg==0.25
PyYAML==5.3
pyzmq==18.1.1
regex==2018.2.21
requests==2.22.0
responses==0.10.9
s3transfer==0.3.2
sacremoses==0.0.38
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
SecretStorage==2.3.1
segtok==1.5.7
sentencepiece==0.1.85
six==1.11.0
sklearn==0.0
smart-open==1.9.0
snowballstemmer==2.0.0
spacy==2.1.9
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlitedict==1.6.0
sqlparse==0.3.0
srsly==1.0.1
tabulate==0.8.6
tensorboard==1.14.0
tensorboardX==2.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
thinc==7.0.8
tokenizers==0.0.11
torch==1.4.0
tqdm==4.42.1
transformers==2.4.1
Unidecode==1.1.1
urllib3==1.24.3
wasabi==0.6.0
wcwidth==0.1.8
Werkzeug==0.15.4
word2number==1.1
wordfreq==2.2.1
wrapt==1.11.2
zipp==2.1.0